feat: #38 전체 다이어리 리스트 불러올 때 그룹이 정상적으로 불러오도록 변경

+++ app/src/main/java/com/km/exchangediary/data/remote/response/JoinedDiaryListResponse.kt
@@ -19,7 +19,7 @@ data class JoinedDiaryListResponseData(
     @SerializedName("now_writer")
     val nowWriter: Long,
     val cover: Int,
-    val group: Int?,
+    val group: Long?,
     @SerializedName("create_at")
     val createdAt: Long,
     @SerializedName("updated_at")
